fix(chat): render live tool activity as log rows
